thesisMIBE/piattaforma/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-emoji-picker==0.0.6
django-js-asset==1.2.2
django-model-utils==4.0.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
